requirements.txt
flask
gevent
RPi.GPIO
lark-parser
